Script used to cut and save geometry from scene
Called from PlayArea.cs script, generates .obj and .mtl files of cutout portions from the original mesh. Also cuts out a separate 'inverted' .obj and .mtl file, which is just a perimeter around the cutout cube. This inverted .obj file is used to take the images used to create the stereo cubemaps for the skybox in the import project.